internal_ui/requirements.txt
streamlit==1.43.0
google-cloud-firestore==2.20.1
pandas==2.2.3
plotly==6.0.0
python-dotenv==1.0.1
